backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Flagged backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiomysql>=0.2.0",
    "fastapi[standard]>=0.115.12",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.41",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
